Fixes to newly broken vortex particle stuff

+++ include/cvortex/libcvtx.h
@@ -86,21 +86,21 @@ CVTX_EXPORT bsv_V3f cvtx_Particle_visc_ind_dvort(
 
 CVTX_EXPORT bsv_V3f cvtx_ParticleArr_ind_vel(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f mes_point,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius);
 
 CVTX_EXPORT bsv_V3f cvtx_ParticleArr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle *induced_particle,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius);
 
 CVTX_EXPORT bsv_V3f cvtx_ParticleArr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle *induced_particle,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius,
@@ -108,27 +108,27 @@ CVTX_EXPORT bsv_V3f cvtx_ParticleArr_visc_ind_dvort(
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_vel(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f *mes_start,
-	const long num_mes,
+	const int num_mes,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius);
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius);
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius,
@@ -152,26 +152,26 @@ CVTX_EXPORT bsv_V3f cvtx_StraightVortFil_ind_dvort(
 
 CVTX_EXPORT bsv_V3f cvtx_StraightVortFilArr_ind_vel(
 	const cvtx_StraightVortFil **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f mes_point);
 
 CVTX_EXPORT bsv_V3f cvtx_StraightVortFilArr_ind_dvort(
 	const cvtx_StraightVortFil **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle *induced_particle);
 
 CVTX_EXPORT void cvtx_StraightVortFilArr_Arr_ind_vel(
 	const cvtx_StraightVortFil **array_start,
-	const long num_filaments,
+	const int num_filaments,
 	const bsv_V3f *mes_start,
-	const long num_mes,
+	const int num_mes,
 	bsv_V3f *result_array);
 
 CVTX_EXPORT void cvtx_StraightVortFilArr_Arr_ind_dvort(
 	const cvtx_StraightVortFil **array_start,
-	const long num_filaments,
+	const int num_filaments,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array);
 
 #endif /* CVTX_LIBCVTX_H */
