+++ src/Particle.c
@@ -122,7 +122,7 @@ CVTX_EXPORT bsv_V3f cvtx_Particle_visc_ind_dvort(
 
 CVTX_EXPORT bsv_V3f cvtx_ParticleArr_ind_vel(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f mes_point,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius)
@@ -144,7 +144,7 @@ CVTX_EXPORT bsv_V3f cvtx_ParticleArr_ind_vel(
 
 CVTX_EXPORT bsv_V3f cvtx_ParticleArr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle *induced_particle,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius)
@@ -166,7 +166,7 @@ CVTX_EXPORT bsv_V3f cvtx_ParticleArr_ind_dvort(
 
 CVTX_EXPORT bsv_V3f cvtx_ParticleArr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle *induced_particle,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius,
@@ -189,9 +189,9 @@ CVTX_EXPORT bsv_V3f cvtx_ParticleArr_visc_ind_dvort(
 
 static void cpu_brute_force_ParticleArr_Arr_ind_vel(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f *mes_start,
-	const long num_mes,
+	const int num_mes,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius)
@@ -208,9 +208,9 @@ static void cpu_brute_force_ParticleArr_Arr_ind_vel(
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_vel(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f *mes_start,
-	const long num_mes,
+	const int num_mes,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius)
@@ -233,9 +233,9 @@ CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_vel(
 
 void cpu_brute_force_ParticleArr_Arr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius)
@@ -252,9 +252,9 @@ void cpu_brute_force_ParticleArr_Arr_ind_dvort(
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius)
@@ -277,9 +277,9 @@ CVTX_EXPORT void cvtx_ParticleArr_Arr_ind_dvort(
 
 void cpu_brute_force_ParticleArr_Arr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius,
@@ -297,9 +297,9 @@ void cpu_brute_force_ParticleArr_Arr_visc_ind_dvort(
 
 CVTX_EXPORT void cvtx_ParticleArr_Arr_visc_ind_dvort(
 	const cvtx_Particle **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array,
 	const cvtx_VortFunc *kernel,
 	float regularisation_radius,
