+++ src/StraightVortFil.c
@@ -77,7 +77,7 @@ CVTX_EXPORT bsv_V3f cvtx_StraightVortFil_ind_dvort(
 
 CVTX_EXPORT bsv_V3f cvtx_StraightVortFilArr_ind_vel(
 	const cvtx_StraightVortFil **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f mes_point) 
 {
 	assert(num_particles >= 0);
@@ -98,7 +98,7 @@ CVTX_EXPORT bsv_V3f cvtx_StraightVortFilArr_ind_vel(
 
 CVTX_EXPORT bsv_V3f cvtx_StraightVortFilArr_ind_dvort(
 	const cvtx_StraightVortFil **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle *induced_particle) 
 {
 	assert(num_particles >= 0);
@@ -119,9 +119,9 @@ CVTX_EXPORT bsv_V3f cvtx_StraightVortFilArr_ind_dvort(
 
 void cpu_brute_force_StraightVortFilArr_Arr_ind_vel(
 	const cvtx_StraightVortFil **array_start,
-	const long num_particles,
+	const int num_particles,
 	const bsv_V3f *mes_start,
-	const long num_mes,
+	const int num_mes,
 	bsv_V3f *result_array) 
 {
 	long i;
@@ -135,9 +135,9 @@ void cpu_brute_force_StraightVortFilArr_Arr_ind_vel(
 
 void cpu_brute_force_StraightVortFilArr_Arr_ind_dvort(
 	const cvtx_StraightVortFil **array_start,
-	const long num_particles,
+	const int num_particles,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array) 
 {
 	long i;
@@ -151,9 +151,9 @@ void cpu_brute_force_StraightVortFilArr_Arr_ind_dvort(
 
 CVTX_EXPORT void cvtx_StraightVortFilArr_Arr_ind_vel(
 	const cvtx_StraightVortFil **array_start,
-	const long num_filaments,
+	const int num_filaments,
 	const bsv_V3f *mes_start,
-	const long num_mes,
+	const int num_mes,
 	bsv_V3f *result_array)
 {
 #ifdef CVTX_USING_OPENCL
@@ -173,9 +173,9 @@ CVTX_EXPORT void cvtx_StraightVortFilArr_Arr_ind_vel(
 
 CVTX_EXPORT void cvtx_StraightVortFilArr_Arr_ind_dvort(
 	const cvtx_StraightVortFil **array_start,
-	const long num_fil,
+	const int num_fil,
 	const cvtx_Particle **induced_start,
-	const long num_induced,
+	const int num_induced,
 	bsv_V3f *result_array)
 {
 #ifdef CVTX_USING_OPENCL
